+++ R/natural_join.R
@@ -160,21 +160,23 @@ to_sql.relop_natural_join <- function (x,
   using <- calc_used_relop_natural_join(x,
                                         using=using)
   c1 <- intersect(using, column_names(x$source[[1]]))
-  subsqla <- to_sql(x$source[[1]],
-                    db = db,
-                    source_limit = source_limit,
-                    indent_level = indent_level + 1,
-                    tnum = tnum,
-                    append_cr = FALSE,
-                    using = c1)
+  subsqla_list <- to_sql(x$source[[1]],
+                         db = db,
+                         source_limit = source_limit,
+                         indent_level = indent_level + 1,
+                         tnum = tnum,
+                         append_cr = FALSE,
+                         using = c1)
+  subsqla <- subsqla_list[[length(subsqla_list)]]
   c2 <- intersect(using, column_names(x$source[[2]]))
-  subsqlb <- to_sql(x$source[[2]],
-                    db = db,
-                    source_limit = source_limit,
-                    indent_level = indent_level + 1,
-                    tnum = tnum,
-                    append_cr = FALSE,
-                    using = c2)
+  subsqlb_list <- to_sql(x$source[[2]],
+                         db = db,
+                         source_limit = source_limit,
+                         indent_level = indent_level + 1,
+                         tnum = tnum,
+                         append_cr = FALSE,
+                         using = c2)
+  subsqlb <- subsqlb_list[[length(subsqlb_list)]]
   taba <- tnum()
   tabaq <- quote_identifier(db, taba)
   tabb <- tnum()
@@ -232,5 +234,7 @@ to_sql.relop_natural_join <- function (x,
   if(append_cr) {
     q <- paste0(q, "\n")
   }
-  q
+  c(subsqla_list[-length(subsqla_list)],
+    subsqlb_list[-length(subsqlb_list)],
+    q)
 }