+++ R/package.R
@@ -18,5 +18,5 @@ NULL
 
 
 
-#' @importFrom wrapr %.>% := mk_tmp_name_source wrapr_function
+#' @importFrom wrapr %.>% := mk_tmp_name_source wrapr_function pipe_step
 NULL