+++ R/select_columns.R
@@ -114,13 +114,14 @@ to_sql.relop_select_columns <- function (x,
   }
   using <- calc_using_relop_select_columns(x,
                                            using = using)
-  subsql <- to_sql(x$source[[1]],
-                   db = db,
-                   source_limit = source_limit,
-                   indent_level = indent_level + 1,
-                   tnum = tnum,
-                   append_cr = FALSE,
-                   using = using)
+  subsql_list <- to_sql(x$source[[1]],
+                        db = db,
+                        source_limit = source_limit,
+                        indent_level = indent_level + 1,
+                        tnum = tnum,
+                        append_cr = FALSE,
+                        using = using)
+  subsql <- subsql_list[[length(subsql_list)]]
   cols <- vapply(x$columns,
                  function(ci) {
                    quote_identifier(db, ci)
@@ -136,7 +137,7 @@ to_sql.relop_select_columns <- function (x,
   if(append_cr) {
     q <- paste0(q, "\n")
   }
-  q
+  c(subsql_list[-length(subsql_list)], q)
 }