+++ R/theta_join.R
@@ -353,20 +353,22 @@ to_sql.relop_theta_join <- function (x,
                                       using=using)
   c1 <- intersect(using, column_names(x$source[[1]]))
   c2 <- intersect(using, column_names(x$source[[2]]))
-  subsqla <- to_sql(x$source[[1]],
-                    db = db,
-                    source_limit = source_limit,
-                    indent_level = indent_level + 1,
-                    tnum = tnum,
-                    append_cr = FALSE,
-                    using = c1)
-  subsqlb <- to_sql(x$source[[2]],
-                    db = db,
-                    source_limit = source_limit,
-                    indent_level = indent_level + 1,
-                    tnum = tnum,
-                    append_cr = FALSE,
-                    using = c2)
+  subsqla_list <- to_sql(x$source[[1]],
+                         db = db,
+                         source_limit = source_limit,
+                         indent_level = indent_level + 1,
+                         tnum = tnum,
+                         append_cr = FALSE,
+                         using = c1)
+  subsqla <- subsqla_list[[length(subsqla_list)]]
+  subsqlb_list <- to_sql(x$source[[2]],
+                         db = db,
+                         source_limit = source_limit,
+                         indent_level = indent_level + 1,
+                         tnum = tnum,
+                         append_cr = FALSE,
+                         using = c2)
+  subsqlb <- subsqlb_list[[length(subsqlb_list)]]
   taba <- tnum()
   tabb <- tnum()
   bterms <- setdiff(column_names(x$source[[1]]),
@@ -401,5 +403,7 @@ to_sql.relop_theta_join <- function (x,
   if(append_cr) {
     q <- paste0(q, "\n")
   }
-  q
+  c(subsqla_list[-length(subsqla_list)],
+    subsqlb_list[-length(subsqlb_list)],
+    q)
 }